feat: primary selection threshold
The primary_threshold_ms variable defines the time interval, in
milliseconds, that controls how often the program updates its stored
value of the Linux primary selection. In the Linux environment, the
primary selection is a mechanism that automatically updates to reflect
the current highlighted text or object, typically updating with every
mouse movement.